+++ GRDB/Core/Row.swift
@@ -156,7 +156,7 @@ extension Row {
     /// righmost column.
     public func value(atIndex index: Int) -> DatabaseValueConvertible? {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
-        return impl.databaseValue(atUncheckedIndex: index).value()
+        return impl.databaseValue(atUncheckedIndex: index).storage.value
     }
     
     /// Returns the value at given index, converted to the requested type.
@@ -169,7 +169,7 @@ extension Row {
     /// fail, a fatal error is raised.
     public func value<Value: DatabaseValueConvertible>(atIndex index: Int) -> Value? {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
-        return impl.databaseValue(atUncheckedIndex: index).value()
+        return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
     }
     
     /// Returns the value at given index, converted to the requested type.
@@ -187,7 +187,7 @@ extension Row {
     public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(atIndex index: Int) -> Value? {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
         guard let sqliteStatement = sqliteStatement else {
-            return impl.databaseValue(atUncheckedIndex: index).value()
+            return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -201,7 +201,7 @@ extension Row {
     /// SQLite value can not be converted to `Value`.
     public func value<Value: DatabaseValueConvertible>(atIndex index: Int) -> Value {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
-        return impl.databaseValue(atUncheckedIndex: index).value()
+        return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
     }
     
     /// Returns the value at given index, converted to the requested type.
@@ -218,7 +218,7 @@ extension Row {
     public func value<Value: DatabaseValueConvertible & StatementColumnConvertible>(atIndex index: Int) -> Value {
         GRDBPrecondition(index >= 0 && index < count, "row index out of range")
         guard let sqliteStatement = sqliteStatement else {
-            return impl.databaseValue(atUncheckedIndex: index).value()
+            return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -241,7 +241,7 @@ extension Row {
         guard let index = impl.index(ofColumn: columnName) else {
             return nil
         }
-        return impl.databaseValue(atUncheckedIndex: index).value()
+        return impl.databaseValue(atUncheckedIndex: index).storage.value
     }
     
     /// Returns the value at given column, converted to the requested type.
@@ -256,7 +256,7 @@ extension Row {
         guard let index = impl.index(ofColumn: columnName) else {
             return nil
         }
-        return impl.databaseValue(atUncheckedIndex: index).value()
+        return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
     }
     
     /// Returns the value at given column, converted to the requested type.
@@ -276,7 +276,7 @@ extension Row {
             return nil
         }
         guard let sqliteStatement = sqliteStatement else {
-            return impl.databaseValue(atUncheckedIndex: index).value()
+            return Value.convertOptional(from: impl.databaseValue(atUncheckedIndex: index))
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -295,7 +295,7 @@ extension Row {
             // Programmer error
             fatalError("no such column: \(columnName)")
         }
-        return impl.databaseValue(atUncheckedIndex: index).value()
+        return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
     }
     
     /// Returns the value at given column, converted to the requested type.
@@ -317,7 +317,7 @@ extension Row {
             fatalError("no such column: \(columnName)")
         }
         guard let sqliteStatement = sqliteStatement else {
-            return impl.databaseValue(atUncheckedIndex: index).value()
+            return Value.convert(from: impl.databaseValue(atUncheckedIndex: index))
         }
         return Row.statementColumnConvertible(atUncheckedIndex: index, in: sqliteStatement)
     }
@@ -888,7 +888,7 @@ private struct ArrayRowImpl : RowImpl {
     }
     
     func dataNoCopy(atUncheckedIndex index:Int) -> Data? {
-        return databaseValue(atUncheckedIndex: index).value()
+        return Data.convertOptional(from: databaseValue(atUncheckedIndex: index))
     }
     
     func databaseValue(atUncheckedIndex index: Int) -> DatabaseValue {
@@ -940,7 +940,7 @@ private struct StatementCopyRowImpl : RowImpl {
     }
     
     func dataNoCopy(atUncheckedIndex index:Int) -> Data? {
-        return databaseValue(atUncheckedIndex: index).value()
+        return Data.convertOptional(from: databaseValue(atUncheckedIndex: index))
     }
     
     func databaseValue(atUncheckedIndex index: Int) -> DatabaseValue {