+++ GRDB/Legacy/Fixits-0.102.0.swift
@@ -2,3 +2,8 @@ extension Database {
     @available(*, unavailable, renamed:"inTransaction")
     public func writeInTransaction(_ kind: Database.TransactionKind? = nil, _ block: (Database) throws -> Database.TransactionCompletion) throws { }
 }
+
+extension DatabaseValue {
+    @available(*, unavailable, message:"DatabaseSequence has been replaced by DatabaseCursor.")
+    public func value() -> Any { preconditionFailure() }
+}